+++ Sources/MapboxMaps/Style/CustomSources/CustomRasterSource.swift
@@ -19,10 +19,15 @@ public struct CustomRasterSource: Source, Equatable {
     @_documentation(visibility: public)
     public let options: CustomRasterSourceOptions?
 
+    /// When a set of tiles for a current zoom level is being rendered and some of the ideal tiles that cover the screen are not yet loaded, parent tiles could be used instead. Note that this might introduce unwanted rendering side-effects, especially for raster tiles that are overscaled multiple times. This property sets the maximum limit for how much a parent tile can be overscaled.
     @_documentation(visibility: public)
-    public init(id: String, options: CustomRasterSourceOptions) {
+    public var maxOverscaleFactorForParentTiles: UInt8?
+
+    @_documentation(visibility: public)
+    public init(id: String, options: CustomRasterSourceOptions, maxOverscaleFactorForParentTiles: UInt8? = nil) {
         self.id = id
         self.options = options
+        self.maxOverscaleFactorForParentTiles = maxOverscaleFactorForParentTiles
     }
 }
 
@@ -30,6 +35,7 @@ extension CustomRasterSource {
     enum CodingKeys: String, CodingKey {
         case id
         case type
+        case maxOverscaleFactorForParentTiles = "max-overscale-factor-for-parent-tiles"
     }
 
     /// Init from a decoder, note that the CustomRasterSourceOptions are not decodable and need to be set separately
@@ -37,12 +43,25 @@ extension CustomRasterSource {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         id = try container.decode(String.self, forKey: .id)
         options = nil
+        maxOverscaleFactorForParentTiles = try container.decodeIfPresent(UInt8.self, forKey: .maxOverscaleFactorForParentTiles)
     }
 
     /// Encode, note that options will not be included
     public func encode(to encoder: Encoder) throws {
         var container = encoder.container(keyedBy: CodingKeys.self)
-        try encodeNonVolatile(to: encoder, into: &container)
+
+        if encoder.userInfo[.volatilePropertiesOnly] as? Bool == true {
+            try encodeVolatile(to: encoder, into: &container)
+        } else if encoder.userInfo[.nonVolatilePropertiesOnly] as? Bool == true {
+            try encodeNonVolatile(to: encoder, into: &container)
+        } else {
+            try encodeVolatile(to: encoder, into: &container)
+            try encodeNonVolatile(to: encoder, into: &container)
+        }
+    }
+
+    private func encodeVolatile(to encoder: Encoder, into container: inout KeyedEncodingContainer<CodingKeys>) throws {
+        try container.encodeIfPresent(maxOverscaleFactorForParentTiles, forKey: .maxOverscaleFactorForParentTiles)
     }
 
     private func encodeNonVolatile(to encoder: Encoder, into container: inout KeyedEncodingContainer<CodingKeys>) throws {